requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
black @ git+https://github.com/psf/black@de65741b8d49d78fa2675ef79b799cd35e92e7c1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
discord==2.2.2
discord.py==2.2.2
distlib==0.3.6
filelock==3.12.0
frozenlist==1.3.3
identify==2.5.22
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pre-commit==3.2.2
PyYAML==6.0
tomli==2.0.1
typing_extensions==4.5.0
virtualenv==20.22.0
yarl==1.9.1
python-dotenv~=1.0.0
SQLAlchemy~=2.0.11